Simplify pipeline for picture-in-picture and scene making.
Remove videoscales and videorate within scene creation, but move it to
the predefined sources.

PIP is still volatile when switching between the same camera-source as
in (broadcast (list cam+mic pip)). I think this is because of the
nature of avfvideosrc and needing to be explicit about caps when
linking to the videobox.